streamingccl: refactor completeStreamIngestion into StreamIngestManager
Previously stream ingestion side APIs were located under streamproducer
package which doesn't make sense, this PR creates a new
StreamIngestManager under streamingest to represent ingestion side APIs.

Release note: None
