+++ pkg/ccl/streamingccl/streamingest/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "streamingest",
     srcs = [
         "metrics.go",
+        "stream_ingest_manager.go",
         "stream_ingestion_frontier_processor.go",
         "stream_ingestion_job.go",
         "stream_ingestion_planning.go",
@@ -39,6 +40,7 @@ go_library(
         "//pkg/sql/rowenc",
         "//pkg/sql/rowexec",
         "//pkg/sql/sem/tree",
+        "//pkg/sql/sessiondata",
         "//pkg/sql/types",
         "//pkg/storage",
         "//pkg/streaming",
