+++ pkg/ccl/streamingccl/streamproducer/BUILD.bazel
@@ -29,7 +29,6 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
         "//pkg/sql/types",
         "//pkg/streaming",
         "//pkg/testutils",
