+++ pkg/streaming/api.go
@@ -27,20 +27,17 @@ func (j StreamID) SafeValue() {}
 // InvalidStreamID is the zero value for StreamID corresponding to no stream.
 const InvalidStreamID StreamID = 0
 
-// GetReplicationStreamManagerHook is the hook to get access to the replication API.
+// GetReplicationStreamManagerHook is the hook to get access to the producer side replication APIs.
 // Used by builtin functions to trigger streaming replication.
 var GetReplicationStreamManagerHook func(evalCtx *tree.EvalContext) (ReplicationStreamManager, error)
 
-// ReplicationStreamManager represents a collection of APIs that streaming replication supports.
-type ReplicationStreamManager interface {
-	// CompleteStreamIngestion signals a running stream ingestion job to complete on the consumer side.
-	CompleteStreamIngestion(
-		evalCtx *tree.EvalContext,
-		txn *kv.Txn,
-		streamID StreamID,
-		cutoverTimestamp hlc.Timestamp,
-	) error
+// GetStreamIngestManagerHook is the hook to get access to the ingestion side replication APIs.
+// Used by builtin functions to trigger streaming replication.
+var GetStreamIngestManagerHook func(evalCtx *tree.EvalContext) (StreamIngestManager, error)
 
+// ReplicationStreamManager represents a collection of APIs that streaming replication supports
+// on the production side.
+type ReplicationStreamManager interface {
 	// StartReplicationStream starts a stream replication job for the specified tenant on the producer side.
 	StartReplicationStream(
 		evalCtx *tree.EvalContext,
@@ -56,7 +53,8 @@ type ReplicationStreamManager interface {
 		txn *kv.Txn) (streampb.StreamReplicationStatus, error)
 
 	// StreamPartition starts streaming replication on the producer side for the partition specified
-	// by opaqueSpec which contains serialized streampb.StreamPartitionSpec protocol message.
+	// by opaqueSpec which contains serialized streampb.StreamPartitionSpec protocol message and
+	// returns a value generator which yields events for the specified partition.
 	StreamPartition(
 		evalCtx *tree.EvalContext,
 		streamID StreamID,
@@ -76,10 +74,30 @@ type ReplicationStreamManager interface {
 	) error
 }
 
+// StreamIngestManager represents a collection of APIs that streaming replication supports
+// on the ingestion side.
+type StreamIngestManager interface {
+	// CompleteStreamIngestion signals a running stream ingestion job to complete on the consumer side.
+	CompleteStreamIngestion(
+		evalCtx *tree.EvalContext,
+		txn *kv.Txn,
+		streamID StreamID,
+		cutoverTimestamp hlc.Timestamp,
+	) error
+}
+
 // GetReplicationStreamManager returns a ReplicationStreamManager if a CCL binary is loaded.
 func GetReplicationStreamManager(evalCtx *tree.EvalContext) (ReplicationStreamManager, error) {
 	if GetReplicationStreamManagerHook == nil {
 		return nil, errors.New("replication streaming requires a CCL binary")
 	}
 	return GetReplicationStreamManagerHook(evalCtx)
+}
+
+// GetStreamIngestManager returns a StreamIngestManager if a CCL binary is loaded.
+func GetStreamIngestManager(evalCtx *tree.EvalContext) (StreamIngestManager, error) {
+	if GetReplicationStreamManagerHook == nil {
+		return nil, errors.New("replication streaming requires a CCL binary")
+	}
+	return GetStreamIngestManagerHook(evalCtx)
 }
